pyproject.toml
[project]
name = "ampav-core"
version = "0.0.10"
description = "Core Libraries/Schemas/Utilties for the AMPAV environment"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",   
]
authors = [
    {name = "Brian Wheeler", email="bdwheele@iu.edu"},
]

requires-python = ">=3.12"
dependencies = [
    "pydantic",
    "pyyaml",
    "av",
]

[project.scripts]
ampav_probe_media = "ampav.core.schema.av_metadata:cli_probe_media"
ampav_import_3play_json = "ampav.core.formats.transcript.threeplay:cli_import_threeplay_json"
ampav_import_text_transcript = "ampav.core.formats.transcript.textfile:cli_import_text_transcript"

[project.urls]
Homepage = "https://github.com/AMPAV/ampav-core"
Repository = "https://github.com/AMPAV/ampav-core.git"
Issues = "https://github.com/AMPAV/ampav-core/issues"



[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

